Rename(#18): 패키지 이름 수정
Resolvers: #18
