+++ application/wypl-core/src/main/java/com/wypl/wyplcore/WyplColorApplication.java
@@ -1,13 +1,13 @@
-package com.wypl.wyplcalendar;
+package com.wypl.wyplcore;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class WyplCalendarApplication {
+public class WyplColorApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(WyplCalendarApplication.class, args);
+		SpringApplication.run(WyplColorApplication.class, args);
 	}
 
 }
