+++ application/wypl-core/src/test/java/com/wypl/wyplcore/WyplCalendarApplicationTests.java
@@ -1,4 +1,4 @@
-package com.wypl.wyplcalendar;
+package com.wypl.wyplcore;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
