+++ domain/jpa-common/src/main/java/com/wypl/jpacommon/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.wypl.jpacore;
+package com.wypl.jpacommon;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
