+++ domain/mongo-common/src/main/java/com/wypl/mongocommon/MongoAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.wypl.mongocore;
+package com.wypl.mongocommon;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
